Move connected components to containers folder
