+++ src/components/MovieList/MovieList.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import MovieCard from '../MovieCard/MovieCard';
+import MovieCard from '../../containers/MovieCard/MovieCard';
 import './MovieList.css';
