+++ src/containers/Favorites/Favorites.css
