+++ src/containers/Favorites/Favorites.js
