+++ src/containers/Favorites/Favotites.test.js
