+++ src/containers/Main/Main.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import MovieList from '../MovieList/MovieList';
+import MovieList from '../../components/MovieList/MovieList';
 import { connect } from 'react-redux';
 import './Main.css';
